split utils.ts
